+++ src/environments/environment.ts
@@ -1,6 +1,6 @@
 export const environment = {
   layout: 'index',
-  version: '20.1.4',
+  version: '20.1.5',
   defaultPage: 'docs',
   static: 'https://ngnest.com/static'
 };